Add the ability to remove nodes from memory arena
This commit adds the ability to remove nodes from the memory arena
resp. the Arena struct. This is achieved by adding a layer between the
access to the nodes by a NodeId and the Vec in the Arena struct. The
layer consists of a HashMap which serves as a lookup table and maps a
NodeId to the index of the corresponding node in the arena. If a node is
removed, it is removed from the vector and the place in the vector is filled
with the last element which is moved to the 'gap'. This will prevent a
leakage of memory after a long time in an approach which uses only a
vector.

The downside of this approach that thread safety is dropped as HashMaps
are not thread safe.